feat(analytics): remove new_analytics_page flag and legacy page
Promote the new analytics page to be the AnalyticsPage, drop the
feature flag, and delete the legacy page along with the components and
SWR hooks it exclusively used.
