+++ sdks/js/src/types.ts
@@ -764,7 +764,6 @@ const WhitelistableFeaturesSchema = FlexibleEnumSchema<
   | "run_tools_from_prompt"
   | "usage_data_api"
   | "usage_page_read_only"
-  | "new_analytics_page"
   | "workspace_analytics"
   | "xai_feature"
   | "conversations_slack_notifications"
